Remove two unnecessary int casts, as reported in [forum:84fe63b7b3|forum post 84fe63b7b3].
FossilOrigin-Name: 137cc3e9e8b080678868da34e61f65faf5feb19f146bb7789c659bcdb8f5e950

+++ src/btree.c
@@ -11993,13 +11993,13 @@ int sqlite3BtreeExclusiveLock(Btree *p){
         if( pSchema ){
           for(pE=sqliteHashFirst(&pSchema->tblHash); pE; pE=sqliteHashNext(pE)){
             Table *pTab = (Table *)sqliteHashData(pE);
-            if( pTab->tnum==(int)pgnoRoot ){
+            if( pTab->tnum==pgnoRoot ){
               zObj = pTab->zName;
               zTab = 0;
             }else{
               Index *pIdx;
               for(pIdx=pTab->pIndex; pIdx; pIdx=pIdx->pNext){
-                if( pIdx->tnum==(int)pgnoRoot ){
+                if( pIdx->tnum==pgnoRoot ){
                   zObj = pIdx->zName;
                   zTab = pTab->zName;
                 }
